Passing test for US2

+++ db/migrate/20190405160724_create_snacks.rb
@@ -2,7 +2,7 @@ class CreateSnacks < ActiveRecord::Migration[5.1]
   def change
     create_table :snacks do |t|
       t.string :name
-      t.money  :price
+      t.float  :price
     end
   end
 end
